+++ src/indexing/faiss_backend/index_utils.py
